chore: move utils/game_init to core/game_init, utils/specials to core/specials
